+++ ExpandWorldSize/compatibility/BetterContinents.cs
@@ -7,20 +7,25 @@ namespace ExpandWorldSize;
 public class BetterContinents
 {
   public const string GUID = "BetterContinents";
-  private static Assembly? BetterContinentsAssembly;
+  private static Assembly? Assembly;
   private static FieldInfo? SettingsField;
   private static FieldInfo? IsEnabledField;
+  private static FieldInfo? WorldSizeField;
   public static void Run()
   {
     if (!Chainloader.PluginInfos.TryGetValue(GUID, out var info)) return;
-    BetterContinentsAssembly = info.Instance.GetType().Assembly;
-    var type = BetterContinentsAssembly.GetType("BetterContinents.BetterContinents");
+    Assembly = info.Instance.GetType().Assembly;
+    var type = Assembly.GetType("BetterContinents.BetterContinents");
     if (type == null) return;
     SettingsField = AccessTools.Field(type, "Settings");
     if (SettingsField == null) return;
-    type = BetterContinentsAssembly.GetType("BetterContinents.BetterContinents+BetterContinentsSettings");
+    type = Assembly.GetType("BetterContinents.BetterContinents+BetterContinentsSettings");
     IsEnabledField = AccessTools.Field(type, "EnabledForThisWorld");
     if (IsEnabledField == null) return;
+    type = Assembly.GetType("BetterContinents.BetterContinents");
+    if (type == null) return;
+    WorldSizeField = AccessTools.Field(type, "WorldSize");
+    if (WorldSizeField == null) return;
     EWS.Log.LogInfo("\"Better Continents\" detected. Applying compatibility.");
   }
 
@@ -31,5 +36,10 @@ public class BetterContinents
     var settings = SettingsField.GetValue(null);
     if (settings == null) return false;
     return (bool)IsEnabledField.GetValue(settings);
+  }
+  public static void RefreshSize()
+  {
+    if (WorldSizeField == null) return;
+    WorldSizeField.SetValue(null, Configuration.WorldTotalRadius);
   }
 }